Use toast for showing version updates & make AddVisitCount not a HOC
