--- src/components/version/VersionModal.tsx
@@ -1,8 +1,0 @@
-import { appVersion } from "@/constants/config";
-import VersionAlertDialog from "./VersionAlertDialog";
-
-const VersionModal = () => {
-	return <VersionAlertDialog backendVersion={appVersion} />;
-};
-
-export default VersionModal;
